+++ src/mocks/Mock.js
@@ -12,16 +12,7 @@ const express = require("express");
 
 const tracer = require("../tracer");
 
-const METHODS = {
-  GET: "get",
-  POST: "post",
-  PATCH: "patch",
-  DELETE: "delete",
-  PUT: "put",
-  OPTIONS: "options",
-  HEAD: "head",
-  TRACE: "trace",
-};
+const { HTTP_METHODS } = require("./helpers");
 
 class Mock {
   constructor({ id, routesVariants, getDelay }) {
@@ -38,7 +29,7 @@ class Mock {
         ? routeVariant.method
         : [routeVariant.method];
       methods.forEach((method) => {
-        this._router[METHODS[method]](routeVariant.url, (req, res, next) => {
+        this._router[HTTP_METHODS[method]](routeVariant.url, (req, res, next) => {
           const delay = routeVariant.delay !== null ? routeVariant.delay : this._getDelay();
           if (delay > 0) {
             tracer.verbose(`Applying delay of ${delay}ms to route variant "${this._id}"`);
@@ -49,7 +40,7 @@ class Mock {
             next();
           }
         });
-        this._router[METHODS[method]](
+        this._router[HTTP_METHODS[method]](
           routeVariant.url,
           routeVariant.middleware.bind(routeVariant)
         );