+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 ### Added
 - feat: Validate that routes and mocks files export an array (#101)
+- feat: Validate routes and route variants format (#101)
 ### Changed
 ### Fixed
 ### Removed